requirements.txt
# Settings
pydantic-settings==2.12.0

# Logging
python-json-logger==2.0.7

# Document Parsing
markitdown==0.1.3
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
xlrd==2.0.1
xlwt==1.3.0
pypdf==6.4.0
PyMuPDF==1.25.1
Pillow==11.1.0
pillow-heif==1.1.1
pi-heif==1.1.1
pdf2image==1.17.0
pytesseract==0.3.13
unstructured==0.18.21
unstructured-inference==1.1.2
unstructured-pytesseract==0.3.15
olefile==0.47
chardet==5.2.0

# Database
psycopg2-binary==2.9.10

# HTTP Client
requests==2.32.3

# Table Parsing
camelot-py[cv]==0.11.0
tabula-py==2.9.3

# Deskew and Image Processing
opencv-python-headless==4.10.0.84
deskew==1.5.1
numpy>=2.1.0

# Text Splitters (LangChain)
langchain-text-splitters==0.3.8
